Create upgrade_to_0.8.0.sql
Update of the latest production version of the "adresse" schema.
It implements various objects type :
•	Functions
•	tables, sequences, default values
•	views
•	indexes
•	triggers
•	constrains
•	comments
This upgrade aims to create new tables and setup several functions which automatically update different fields in our tables. It allow to produce assessment view used to export data with a structure expected by the project partners.